refactor: update all balances
