+++ frontend/database/rxdb.ts
@@ -670,13 +670,13 @@ export class RxdbDatabase extends IWalletDatabase {
   private async _tokensPushHandler(items: any[]): Promise<TokenRxdbDocument[]> {
     const arg = items.map(
       (x) =>
-      ({
-        ...x,
-        id_number: x.id_number,
-        updatedAt: Math.floor(Date.now() / 1000),
-        logo: extractValueFromArray(x.logo),
-        index: extractValueFromArray(x.index),
-      } as TokenRxdbDocument),
+        ({
+          ...x,
+          id_number: x.id_number,
+          updatedAt: Math.floor(Date.now() / 1000),
+          logo: extractValueFromArray(x.logo),
+          index: extractValueFromArray(x.index),
+        } as TokenRxdbDocument),
     );
 
     await this.replicaCanister?.pushTokens(arg);
@@ -714,11 +714,11 @@ export class RxdbDatabase extends IWalletDatabase {
           allowance:
             !!s.allowance && !!s.allowance.allowance
               ? [
-                {
-                  allowance: [s.allowance.allowance],
-                  expires_at: [s.allowance.expires_at],
-                },
-              ]
+                  {
+                    allowance: [s.allowance.allowance],
+                    expires_at: [s.allowance.expires_at],
+                  },
+                ]
               : [],
         })),
       })),
@@ -790,7 +790,7 @@ export class RxdbDatabase extends IWalletDatabase {
             deleted: false,
             updatedAt: Date.now(),
             // TODO: fee should not be here
-            fee: ""
+            fee: "",
           })),
         );
       } catch (e) {
