+++ frontend/pages/home/components/ICRC/asset/AddAssetAutomatic.tsx
@@ -90,8 +90,9 @@ const AddAssetAutomatic = ({
                   return (
                     <DropdownMenu.Item
                       key={`net-${idx}`}
-                      className={`flex flex-row text-md justify-start items-center p-3 cursor-pointer ${idx > 0 ? "border-t border-BorderColorLight dark:border-BorderColor" : ""
-                        }`}
+                      className={`flex flex-row text-md justify-start items-center p-3 cursor-pointer ${
+                        idx > 0 ? "border-t border-BorderColorLight dark:border-BorderColor" : ""
+                      }`}
                       onSelect={() => {
                         onSelectNetwork(sa);
                       }}
@@ -152,8 +153,9 @@ const AddAssetAutomatic = ({
                     return (
                       <DropdownMenu.Item
                         key={`net-${idx}`}
-                        className={`flex flex-row text-md justify-start items-center p-3 cursor-pointer ${idx > 0 ? "border-t border-BorderColorLight dark:border-BorderColor" : ""
-                          }`}
+                        className={`flex flex-row text-md justify-start items-center p-3 cursor-pointer ${
+                          idx > 0 ? "border-t border-BorderColorLight dark:border-BorderColor" : ""
+                        }`}
                         onSelect={() => {
                           onSelectToken(newAsset);
                         }}
@@ -193,16 +195,18 @@ const AddAssetAutomatic = ({
         tokenSymbol: "",
         decimal: "",
         shortDecimal: "",
-        subAccounts: [{
-          sub_account_id: "0x0",
-          name: AccountDefaultEnum.Values.Default,
-          amount: "0",
-          currency_amount: "0",
-          address: "",
-          decimal: 0,
-          symbol: "",
-          transaction_fee: ""
-        }],
+        subAccounts: [
+          {
+            sub_account_id: "0x0",
+            name: AccountDefaultEnum.Values.Default,
+            amount: "0",
+            currency_amount: "0",
+            address: "",
+            decimal: 0,
+            symbol: "",
+            transaction_fee: "",
+          },
+        ],
         index: "",
         sortIndex: 999,
         supportedStandards: [],
@@ -225,16 +229,18 @@ const AddAssetAutomatic = ({
       tokenSymbol: "",
       decimal: "",
       shortDecimal: "",
-      subAccounts: [{
-        sub_account_id: "0x0",
-        name: AccountDefaultEnum.Values.Default,
-        amount: "0",
-        currency_amount: "0",
-        address: "",
-        decimal: 0,
-        symbol: "",
-        transaction_fee: "",
-      }],
+      subAccounts: [
+        {
+          sub_account_id: "0x0",
+          name: AccountDefaultEnum.Values.Default,
+          amount: "0",
+          currency_amount: "0",
+          address: "",
+          decimal: 0,
+          symbol: "",
+          transaction_fee: "",
+        },
+      ],
       index: "",
       sortIndex: 999,
       supportedStandards: [],
